change class names and organize code in .dart files
change class names in business_verification_success_view.dart
change class names in business_verification_failure_view.dart
remove MaterialApp in business_verification_success_view.dart
remove MaterialApp in bussiness_verification_failure_view.dart
remove unused imports in .dart files
shift VerificationStartView class to main.dart
remove menu in business_registration_details_view.dart
